LayoutTests/imported/w3c/web-platform-tests/editing/other/delete-in-child-of-html.tentative_designMode=on&method=backspace-expected.txt



FAIL delete in <div> elements after <body> should join them assert_in_array: The <div> elements should be merged value "<head><title>iframe</title></head><body><br></body><div>abc</div><div>ef</div>" not in array ["<head><title>iframe</title></head><body><br></body><div>abcdef</div>", "<head><title>iframe</title></head><body><br></body><div>abcdef<br></div>"]
FAIL delete should merge <div> after <body> into the <body> assert_in_array: The text should be merged value "<head><title>iframe</title></head><body>abc</body><div>ef</div>" not in array ["<head><title>iframe</title></head><body>abcdef</body>", "<head><title>iframe</title></head><body>abcdef<br></body>"]
FAIL delete should merge <div> after <body> into the <div> in the <body> assert_in_array: The <div> elements should be merged value "<head><title>iframe</title></head><body><div>abc</div></body><div>ef</div>" not in array ["<head><title>iframe</title></head><body><div>abcdef</div></body>", "<head><title>iframe</title></head><body><div>abcdef<br></div></body>"]
FAIL delete should merge <div> after <body> into the empty <body> assert_in_array: The <div> element should be merged into the <body> value "<head><title>iframe</title></head><body><br></body><div>bc</div>" not in array ["<head><title>iframe</title></head><body>abc</body>", "<head><title>iframe</title></head><body>abc<br></body>"]
FAIL delete should merge <div> before <body> into the <body> assert_in_array: The text should be merged value "<head><title>iframe</title></head><div>abc</div><body>ef</body>" not in array ["<head><title>iframe</title></head><body>abcdef</body>", "<head><title>iframe</title></head><body>abcdef<br></body>"]
FAIL delete should merge <div> before <body> into the <div> in the <body> assert_in_array: The <div> elements should be merged value "<head><title>iframe</title></head><div>abc</div><body><div>ef</div></body>" not in array ["<head><title>iframe</title></head><body><div>abcdef</div></body>", "<head><title>iframe</title></head><body><div>abcdef<br></div></body>"]
FAIL delete should merge <div> before <body> into the empty <body> assert_in_array: The <div> element should be merged into the <body> value "<head><title>iframe</title></head><div>abc</div><body><br></body>" not in array ["<head><title>iframe</title></head><body>abc</body>", "<head><title>iframe</title></head><body>abc<br></body>"]
FAIL delete from <div> following invisible <head> element shouldn't delete the <head> element assert_equals: The <div> element should be merged into the <body> expected "<head><title>iframe</title></head><div>abc</div><body>def</body>" but got "<head><title>iframe</title></head><div>bc</div><body>def</body>"
FAIL delete from <div> around invisible <head> element should not delete the <head> assert_in_array: The <div> element should be merged into the left <div> without deleting the <head> value "<div>abc</div><head><title>iframe</title></head><div>ef</div><body><br></body>" not in array ["<div>abcdef</div><head><title>iframe</title></head><body><br></body>", "<div>abcdef<br></div><head><title>iframe</title></head><body><br></body>", "<head><title>iframe</title></head><div>abcdef</div><body><br></body>", "<head><title>iframe</title></head><div>abcdef<br></div><body><br></body>"]
FAIL delete from <div> following visible <head> element should be merged with the <div> in the <head> assert_in_array: The <div> element should be merged into the <div> in the <head> value "<head><title>iframe</title><div>abc</div></head><div>ef</div><body><br></body>" not in array ["<head><title>iframe</title><div>abcdef</div></head><body><br></body>", "<head><title>iframe</title><div>abcdef<br></div></head><body><br></body>"]
FAIL delete from <div> following visible <head> element should be merged with the visible <style> in the <head> assert_equals: The <div> element should not be merged with the <style> in the <head> expected "<head><title>iframe</title><style>abc</style></head><div>def</div><body><br></body>" but got "<head><title>iframe</title><style>abc</style></head><div>ef</div><body><br></body>"
FAIL delete from <div> following visible <script> element should be merged with the visible <script> in the <head> assert_equals: The <div> element should not be merged with the <script> in the <head> expected "<head><title>iframe</title><script>// abc</script></head><div>def</div><body><br></body>" but got "<head><title>iframe</title><script>// abc</script></head><div>ef</div><body><br></body>"
